% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{format_episode_name}
\alias{format_episode_name}
\title{format_episode_name}
\usage{
format_episode_name(episode)
}
\arguments{
\item{episode}{}
}
\value{
a data frame that has been cleaned of NA values
}
\description{
format_episode_name
}
