% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{format_season_name}
\alias{format_season_name}
\title{format_season_name}
\usage{
format_season_name(season)
}
\arguments{
\item{season}{}
}
\value{
a data frame that has been cleaned of NA values
}
\description{
format_season_name
}
\examples{
format_season_name(season)
}
