% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{format_show_name}
\alias{format_show_name}
\title{format_show_name}
\usage{
format_show_name(show)
}
\arguments{
\item{show}{}
}
\value{
a dataframe of the shows returned in the get_shows() function, along
with premier date, end date, and the genres
}
\description{
format_show_name
}
\examples{
format_show_name(get_shows("Hello Kitty"))
}
