% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{generate_ratings_plot}
\alias{generate_ratings_plot}
\title{generate_ratings_plot}
\usage{
generate_ratings_plot(all_episodes_df)
}
\arguments{
\item{all_episodes_df}{}
}
\value{
a line plot representing the average rating across the seasons of the show
}
\description{
generate_ratings_plot
}
\examples{
generate_ratings_plot(all_episodes_df)
}
