% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{generate_season_ratings_plot}
\alias{generate_season_ratings_plot}
\title{generate_season_ratings_plot}
\usage{
generate_season_ratings_plot(season_df)
}
\arguments{
\item{season_df}{}
}
\value{
a plot of the rating for each episode in a given season
}
\description{
generate_season_ratings_plot
}
\examples{
generate_season_ratings_plot(season_df)
}
