% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{get_episodes_of_season}
\alias{get_episodes_of_season}
\title{get_episodes_of_season}
\usage{
get_episodes_of_season(season_id)
}
\arguments{
\item{season_id}{}
}
\value{
a data frame of the parsed json
}
\description{
get_episodes_of_season
}
\examples{
get_episodes_of_season(season_id)
}
