% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{get_seasons}
\alias{get_seasons}
\title{get_seasons}
\usage{
get_seasons(show_id)
}
\arguments{
\item{show_id}{}
}
\value{
parsed json file containing information about the seasons
}
\description{
get_seasons
}
\examples{
get_seasons(1505)
}
