% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{get_shows}
\alias{get_shows}
\title{get_shows}
\usage{
get_shows(query)
}
\arguments{
\item{query}{}
}
\value{
a dataframe containing all information on a given television show
}
\description{
get_shows
}
\examples{
get_shows("Hello Kitty")
}
